# openapiv2/laptop_service.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "laptop_service.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "LaptopService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/laptop/create": {
      "post": {
        "operationId": "LaptopService_CreateLaptop",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/techschool_pcbookCreateLaptopResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/techschool_pcbookCreateLaptopRequest"
            }
          }
        ],
        "tags": [
          "LaptopService"
        ]
      }
    },
    "/v1/laptop/rate": {
      "post": {
        "operationId": "LaptopService_RateLaptop",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "$ref": "#/definitions/techschool_pcbookRateLaptopResponse"
                },
                "error": {
                  "$ref": "#/definitions/rpcStatus"
                }
              },
              "title": "Stream result of techschool_pcbookRateLaptopResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": " (streaming inputs)",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/techschool_pcbookRateLaptopRequest"
            }
          }
        ],
        "tags": [
          "LaptopService"
        ]
      }
    },
    "/v1/laptop/search": {
      "get": {
        "operationId": "LaptopService_SearchLaptop",
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "$ref": "#/definitions/techschool_pcbookSearchLaptopResponse"
                },
                "error": {
                  "$ref": "#/definitions/rpcStatus"
                }
              },
              "title": "Stream result of techschool_pcbookSearchLaptopResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "filter.max_price_usd",
            "in": "query",
            "required": false,
            "type": "number",
            "format": "double"
          },
          {
            "name": "filter.min_cpu_cores",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "filter.min_cpu_ghz",
            "in": "query",
            "required": false,
            "type": "number",
            "format": "double"
          },
          {
            "name": "filter.min_ram.value",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "uint64"
          },
          {
            "name": "filter.min_ram.unit",
            "in": "query",
            "required": false,
            "type": "string",
            "enum": [
              "UNKNOWN",
              "BIT",
              "BYTE",
              "KILOBYTE",
              "MEGABYTE",
              "GIGABYTE",
              "TERABYTE"
            ],
            "default": "UNKNOWN"
          }
        ],
        "tags": [
          "LaptopService"
        ]
      }
    },
    "/v1/laptop/upload": {
      "post": {
        "operationId": "LaptopService_UploadImage",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/techschool_pcbookUploadImageResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": " (streaming inputs)",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/techschool_pcbookUploadImageRequest"
            }
          }
        ],
        "tags": [
          "LaptopService"
        ]
      }
    }
  },
  "definitions": {
    "KeyboardLayout": {
      "type": "string",
      "enum": [
        "UNKNOWN",
        "QWERTY",
        "QWERTZ",
        "AZERTY"
      ],
      "default": "UNKNOWN"
    },
    "MemoryUnit": {
      "type": "string",
      "enum": [
        "UNKNOWN",
        "BIT",
        "BYTE",
        "KILOBYTE",
        "MEGABYTE",
        "GIGABYTE",
        "TERABYTE"
      ],
      "default": "UNKNOWN"
    },
    "ScreenPanel": {
      "type": "string",
      "enum": [
        "UNKNOWN",
        "IPS",
        "OLED"
      ],
      "default": "UNKNOWN"
    },
    "ScreenResolution": {
      "type": "object",
      "properties": {
        "width": {
          "type": "integer",
          "format": "int64"
        },
        "height": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "StorageDriver": {
      "type": "string",
      "enum": [
        "UNKNOWN",
        "HDD",
        "SSD"
      ],
      "default": "UNKNOWN"
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "techschool_pcbookCPU": {
      "type": "object",
      "properties": {
        "brand": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "number_cores": {
          "type": "integer",
          "format": "int64"
        },
        "number_threads": {
          "type": "integer",
          "format": "int64"
        },
        "min_ghz": {
          "type": "number",
          "format": "double"
        },
        "max_ghz": {
          "type": "number",
          "format": "double"
        }
      }
    },
    "techschool_pcbookCreateLaptopRequest": {
      "type": "object",
      "properties": {
        "laptop": {
          "$ref": "#/definitions/techschool_pcbookLaptop"
        }
      }
    },
    "techschool_pcbookCreateLaptopResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        }
      }
    },
    "techschool_pcbookFilter": {
      "type": "object",
      "properties": {
        "max_price_usd": {
          "type": "number",
          "format": "double"
        },
        "min_cpu_cores": {
          "type": "integer",
          "format": "int64"
        },
        "min_cpu_ghz": {
          "type": "number",
          "format": "double"
        },
        "min_ram": {
          "$ref": "#/definitions/techschool_pcbookMemory"
        }
      }
    },
    "techschool_pcbookGPU": {
      "type": "object",
      "properties": {
        "brand": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "min_ghz": {
          "type": "number",
          "format": "double"
        },
        "max_ghz": {
          "type": "number",
          "format": "double"
        },
        "memory": {
          "$ref": "#/definitions/techschool_pcbookMemory"
        }
      }
    },
    "techschool_pcbookImageInfo": {
      "type": "object",
      "properties": {
        "laptop_id": {
          "type": "string"
        },
        "image_type": {
          "type": "string"
        }
      }
    },
    "techschool_pcbookKeyboard": {
      "type": "object",
      "properties": {
        "layout": {
          "$ref": "#/definitions/KeyboardLayout"
        },
        "backlit": {
          "type": "boolean"
        }
      }
    },
    "techschool_pcbookLaptop": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "brand": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "cpu": {
          "$ref": "#/definitions/techschool_pcbookCPU"
        },
        "ram": {
          "$ref": "#/definitions/techschool_pcbookMemory"
        },
        "gpus": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/techschool_pcbookGPU"
          }
        },
        "storages": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/techschool_pcbookStorage"
          }
        },
        "screen": {
          "$ref": "#/definitions/techschool_pcbookScreen"
        },
        "keyboard": {
          "$ref": "#/definitions/techschool_pcbookKeyboard"
        },
        "weight_kg": {
          "type": "number",
          "format": "double"
        },
        "weight_lb": {
          "type": "number",
          "format": "double"
        },
        "price_usd": {
          "type": "number",
          "format": "double"
        },
        "release_year": {
          "type": "integer",
          "format": "int64"
        },
        "updated_at": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "techschool_pcbookMemory": {
      "type": "object",
      "properties": {
        "value": {
          "type": "string",
          "format": "uint64"
        },
        "unit": {
          "$ref": "#/definitions/MemoryUnit"
        }
      }
    },
    "techschool_pcbookRateLaptopRequest": {
      "type": "object",
      "properties": {
        "laptop_id": {
          "type": "string"
        },
        "score": {
          "type": "number",
          "format": "double"
        }
      }
    },
    "techschool_pcbookRateLaptopResponse": {
      "type": "object",
      "properties": {
        "laptop_id": {
          "type": "string"
        },
        "rated_count": {
          "type": "integer",
          "format": "int64"
        },
        "average_score": {
          "type": "number",
          "format": "double"
        }
      }
    },
    "techschool_pcbookScreen": {
      "type": "object",
      "properties": {
        "size_inch": {
          "type": "number",
          "format": "float"
        },
        "resolution": {
          "$ref": "#/definitions/ScreenResolution"
        },
        "panel": {
          "$ref": "#/definitions/ScreenPanel"
        },
        "multitouch": {
          "type": "boolean"
        }
      }
    },
    "techschool_pcbookSearchLaptopResponse": {
      "type": "object",
      "properties": {
        "laptop": {
          "$ref": "#/definitions/techschool_pcbookLaptop"
        }
      }
    },
    "techschool_pcbookStorage": {
      "type": "object",
      "properties": {
        "driver": {
          "$ref": "#/definitions/StorageDriver"
        },
        "memory": {
          "$ref": "#/definitions/techschool_pcbookMemory"
        }
      }
    },
    "techschool_pcbookUploadImageRequest": {
      "type": "object",
      "properties": {
        "info": {
          "$ref": "#/definitions/techschool_pcbookImageInfo"
        },
        "chunk_data": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "techschool_pcbookUploadImageResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "size": {
          "type": "integer",
          "format": "int64"
        }
      }
    }
  }
}